Restructure react routes to separate module in backend
